# packages/api-spec/specs/openapi.yaml
openapi: 3.0.3
info:
  title: SoulLink Companion API
  description: API for the SoulLink Companion application
  version: 1.0.0
  contact:
    name: API Support
    email: support@soullink.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5001/api/v1
    description: Development server
  - url: https://api.soullink.com/v1
    description: Production server
paths:
...
  /users:
    post:
      summary: Create a new user
      description: Creates a new user with a username
      operationId: createUser
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get user by ID
      description: Returns a user by their ID
      operationId: getUserById
      tags:
        - Users
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      summary: Update user
      description: Update a user's username
      operationId: updateUser
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    Error:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required:
            - message
          properties:
            message:
              type: string
              description: Error message
            code:
              type: string
              description: Error code
            details:
              type: object
              description: Additional error details
      example:
        success: false
        error:
          message: "Resource not found"
          code: "NOT_FOUND"
    User:
      type: object
      required:
        - id
        - username
        - createdAt
      properties:
        id:
          type: string
          description: Unique user identifier
        username:
          type: string
          minLength: 2
          maxLength: 50
          description: User's display name
        createdAt:
          type: string
          format: date-time
          description: When the user was created
      example:
        id: "123e4567-e89b-12d3-a456-426614174000"
        username: "Player1"
        createdAt: "2025-06-20T09:00:00.000Z"
    CreateUserRequest:
      type: object
      required:
        - username
      properties:
        username:
          type: string
          minLength: 2
          maxLength: 50
          description: User's display name
      example:
        username: "Player1"
    CreateUserResponse:
      type: object
      required:
        - user
      properties:
        user:
          $ref: '#/components/schemas/User'
      example:
        user:
          id: "123e4567-e89b-12d3-a456-426614174000"
          username: "Player1"
    GetUserResponse:
      type: object
      required:
        - user
      properties:
        user:
          $ref: '#/components/schemas/User'
      example:
        user:
          id: "123e4567-e89b-12d3-a456-426614174000"
          username: "Player1"
          createdAt: "2025-06-20T09:00:00.000Z"
    UpdateUserRequest:
      type: object
      required:
        - username
      properties:
        username:
          type: string
          minLength: 2
          maxLength: 50
          description: New username for the user
      example:
        username: "NewPlayerName"
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...